+++ docs/dev-dist/sw.js
@@ -82,7 +82,7 @@ define(['./workbox-8027588d'], (function (workbox) { 'use strict';
     "revision": "3ca0b8505b4bec776b69afdba2768812"
   }, {
     "url": "index.html",
-    "revision": "0.tqkuompg98"
+    "revision": "0.h1utev5fhio"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("index.html"), {