+++ packages/backan/CHANGELOG.md
@@ -1,5 +1,14 @@
 # backan
 
+## 0.1.5
+
+### Patch Changes
+
+- Add client package, add more examples, up deps
+
+- Updated dependencies []:
+  - @backan/core@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes