+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backan/core
 
+## 0.1.5
+
+### Patch Changes
+
+- Add client package, add more examples, up deps
+
 ## 0.1.4
 
 ### Patch Changes